+++ frontend/mms-Admin/src/components/DashboardContainer/DashboardContainer.jsx
@@ -1,4 +1,4 @@
-import React, { Suspense, useEffect } from "react";
+import React, { Suspense } from "react";
 import PropTypes from "prop-types";
 import cx from "classnames";
 import styles from "./DashboardContainer.module.scss";
@@ -7,30 +7,9 @@ import DashboardHeader from "../DashboardHeader/DashboardHeader";
 import DashboardSideBar from "../DashboardSideBar/DashboardSideBar";
 import DashboardChildren from "@/components/DashboardChildren/DashboardChildren";
 import Loader from "@/components/Loader/Loader";
-import { refreshTokenInterval } from "@/utils/auth";
-import { refreshAccessToken } from "@/redux/Auth/AuthSlice";
-import { useDispatch, useSelector } from "react-redux";
-import { useNavigate } from "react-router-dom";
 
 function DashboardContainer(props) {
   const { children } = props;
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const authError = useSelector((state) => state?.auth?.error?.failed) || false;
-
-  useEffect(() => {
-    const refreshInterval = setInterval(() => {
-      dispatch(refreshAccessToken());
-    }, refreshTokenInterval());
-
-    if (authError) {
-      clearInterval(refreshInterval);
-      navigate("/login");
-    }
-    return () => {
-      clearInterval(refreshInterval);
-    };
-  }, [authError, dispatch, navigate]);
 
   return (
     <div className={cx(styles.dashboardContainer, "flexCol")}>
